Factory Sense/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
twilio>=9.0.0
python-dotenv>=1.0.0
requests>=2.31.0
